Any changes needed from other branches and/or trunk will have to
be cherry-picked.

Merged revisions 184831,184905,184972,184974 via svnmerge from 
svn+ssh://gcc.gnu.org/svn/gcc/branches/google/main

................
  r184831 | aaw | 2012-03-02 16:32:05 -0500 (Fri, 02 Mar 2012) | 14 lines
  
  Merged r184824 from google/integration.
  
  Add an x86_64-grtev3-linux-gnu target.
          
  gcc/ChangeLog.google-integration
         
  	* config.gcc: Add *-grtev3-* target support.
  	* config/i386/linux.h: Comment fixup.
  	* config/i386/linux64.h: Comment fixup.
  	* config/linux-grtev3.h: New file.
  
  contrib/ChangeLog.google-integration
  	* testsuite-management/x86_64-grtev3-linux-gnu.xfail: New file.
................
  r184905 | davidxl | 2012-03-04 17:09:41 -0500 (Sun, 04 Mar 2012) | 31 lines
  
  Merged revisions 184370,184482,184682 via svnmerge from 
  svn+ssh://gcc.gnu.org/svn/gcc/branches/google/integration
  
  ........
    r184370 | dougkwan | 2012-02-17 22:12:40 -0800 (Fri, 17 Feb 2012) | 12 lines
    
    2012-02-17   Doug Kwan  <dougkwan@google.com>
    
    	* gcc/config.gcc (powerpc*-*-linux): Pull in GRTEv2 spec changes if
    	target matches *-grtev2-*.
    	* gcc/config/rs6000/linux64.h (GLIB_DYNAMIC_LINKER{32,64}): Add
    	runtime root prefix to glibc's dynamic linker.
    	* gcc/config/rs6000/linux-grtev2.h: New file.
    	* gcc/config/rs6000/sysv4.h (GLIB_DYNAMIC_LINKER): Add
    	runtime root prefix to glibc's dynamic linker.
    	(LINUX_GRTE_EXTRA_SPECS): Define to be empty if no definition found.
    	(SUBTARGET_EXTRA_SPECS): Include LINUX_GRTE_EXTRA_SPECS.
  ........
    r184482 | aaw | 2012-02-22 07:20:13 -0800 (Wed, 22 Feb 2012) | 4 lines
    
    2012-02-22  Ollie Wild  <aaw@google.com>
    
           * testsuite-management/x86_64-unknown-linux-gnu.xfail: Add
           gcc.c-torture/execute/vshuf-v16qi.c failure.
  ........
    r184682 | singhai | 2012-02-29 16:56:21 -0800 (Wed, 29 Feb 2012) | 4 lines
    
    Merged revisions 183499-183500,183502-183503,183507,183509-183510,183512-183516,183518-183519,183521-183522,183524,183527-183542,183547,183549,183552,183555-183557,183559-183561,183566-183567,183570,183577,183579-183585,183588,183591,183593,183598-183599,183601-183611,183613-183616,183620-183622,183624-183625,183627-183630,183632-183635,183637-183638,183641-183644,183648,183650-183652,183655,183660-183661,183663,183667-183670,183674-183688,183692-183694,183696-183708,183710-183711,183713-183715,183717-183726,183728,183730,183732-183733,183737,183739,183741,183743-183745,183749,183751-183759,183763,183765-183766,183768-183771,183773-183775,183782,183787-183788,183791-183794,183796,183798,183800,183802,183806-183808,183810-183812,183814-183816,183821,183825-183826,183829-183830,183832-183833,183836,183840-183841,183843,183846-183849,183851-183853,183859-183860,183862,183865,183867,183869-183870,183873,183875,183878,183887,183889,183891,183894-183895,183900,183902,183904-183918,183920,183925-183926,183928,183932-183934,183936-183937,183940-183945,183951,183953-183956,183962,183964,183966,183968,183970-183972,183974,183976-183985,183987,183991,183993-183995,183997-184001,184003-184007,184009-184011,184015-184017,184021-184028,184032-184037,184041-184043,184045,184065,184075,184079-184080,184083,184086,184088-184089,184092,184096-184100,184102-184103,184107-184108,184110-184113,184115-184117,184121-184124,184126-184130,184134-184145,184150-184152,184155-184161,184165,184167-184169,184171,184174,184177,184180-184181,184185,184187-184188,184192-184193,184196,184198-184203,184210-184212,184214,184217-184224,184227,184229,184232-184234,184236-184238,184245-184246,184250-184255,184257-184259,184264-184265,184269,184272,184274,184276-184277,184279-184282,184284,184286-184287,184289,184291,184293,184297-184303,184305-184306,184310,184313-184314,184316-184320,184322,184327-184343,184345,184347-184353,184355-184357,184362-184365,184367,184372-184373,184377,184379-184381,184385,184388,184390-184394,184396,184398-184406,184409,184414,184416-184417,184421-184422,184426,184428-184430,184433-184437,184439-184440,184442,184445-184448,184451,184453,184458,184461-184464,184466-184467,184471-184475,184477-184478,184480-184481,184485,184488,184490,184500-184501,184503,184506-184509,184511,184519,184522,184526-184527,184530-184531,184540,184542-184549,184552,184555,184557-184560,184564,184568-184569,184571-184572,184576,184581-184585,184589-184590,184592-184594,184596-184607,184613-184630,184632-184634,184636,184638-184642,184645,184649,184651-184654,184656-184662,184664-184667 via svnmerge from 
    svn+ssh://gcc.gnu.org/svn/gcc/trunk
  ........
................
  r184972 | aaw | 2012-03-05 23:01:04 -0500 (Mon, 05 Mar 2012) | 5 lines
  
  Update x86_64-grtev3-linux-gnu test manifest.
  
  contrib/
  	* testsuite-management/x86_64-grtev3-linux-gnu.xfail: Updated to
  	reflect current x86_64-unknown-linux-gnu.xfail file.
................
  r184974 | davidxl | 2012-03-06 00:55:36 -0500 (Tue, 06 Mar 2012) | 25 lines
  
  Merged revisions 184824,184851 via svnmerge from 
  svn+ssh://gcc.gnu.org/svn/gcc/branches/google/integration
  
  ........
    r184824 | aaw | 2012-03-02 12:52:24 -0800 (Fri, 02 Mar 2012) | 12 lines
    
    Add an x86_64-grtev3-linux-gnu target.
    
    gcc/ChangeLog.google-integration
    
    	* config.gcc: Add *-grtev3-* target support.
    	* config/i386/linux.h: Comment fixup.
    	* config/i386/linux64.h: Comment fixup.
    	* config/linux-grtev3.h: New file.
    
    contrib/ChangeLog.google-integration
    
    	* testsuite-management/x86_64-grtev3-linux-gnu.xfail: New file.
  ........
    r184851 | singhai | 2012-03-02 20:42:30 -0800 (Fri, 02 Mar 2012) | 4 lines
    
    Merged revisions 184668-184669,184672,184674-184676,184680-184681,184684,184686,184688,184692,184739,184743,184746,184748,184750-184757,184760-184761,184763-184766,184770,184772-184776 via svnmerge from 
    svn+ssh://gcc.gnu.org/svn/gcc/trunk
  ........
................



git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/google/gcc-4_7@185068 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libgo/go/os/exec_windows.go
@@ -11,6 +11,8 @@ import (
 	"unsafe"
 )
 
+// Wait waits for the Process to exit or stop, and then returns a
+// Waitmsg describing its status and an error, if any.
 func (p *Process) Wait(options int) (w *Waitmsg, err error) {
 	s, e := syscall.WaitForSingleObject(syscall.Handle(p.handle), syscall.INFINITE)
 	switch s {
@@ -27,7 +29,7 @@ func (p *Process) Wait(options int) (w *Waitmsg, err error) {
 		return nil, NewSyscallError("GetExitCodeProcess", e)
 	}
 	p.done = true
-	return &Waitmsg{p.Pid, syscall.WaitStatus{s, ec}, new(syscall.Rusage)}, nil
+	return &Waitmsg{p.Pid, syscall.WaitStatus{Status: s, ExitCode: ec}, new(syscall.Rusage)}, nil
 }
 
 // Signal sends a signal to the Process.
@@ -35,36 +37,36 @@ func (p *Process) Signal(sig Signal) error {
 	if p.done {
 		return errors.New("os: process already finished")
 	}
-	switch sig.(UnixSignal) {
-	case SIGKILL:
+	if us, ok := sig.(UnixSignal); ok && us == syscall.SIGKILL {
 		e := syscall.TerminateProcess(syscall.Handle(p.handle), 1)
 		return NewSyscallError("TerminateProcess", e)
 	}
 	return syscall.Errno(syscall.EWINDOWS)
 }
 
+// Release releases any resources associated with the Process.
 func (p *Process) Release() error {
-	if p.handle == -1 {
+	if p.handle == uintptr(syscall.InvalidHandle) {
 		return EINVAL
 	}
 	e := syscall.CloseHandle(syscall.Handle(p.handle))
 	if e != nil {
 		return NewSyscallError("CloseHandle", e)
 	}
-	p.handle = -1
+	p.handle = uintptr(syscall.InvalidHandle)
 	// no need for a finalizer anymore
 	runtime.SetFinalizer(p, nil)
 	return nil
 }
 
-func FindProcess(pid int) (p *Process, err error) {
+func findProcess(pid int) (p *Process, err error) {
 	const da = syscall.STANDARD_RIGHTS_READ |
 		syscall.PROCESS_QUERY_INFORMATION | syscall.SYNCHRONIZE
 	h, e := syscall.OpenProcess(da, false, uint32(pid))
 	if e != nil {
 		return nil, NewSyscallError("OpenProcess", e)
 	}
-	return newProcess(pid, int(h)), nil
+	return newProcess(pid, uintptr(h)), nil
 }
 
 func init() {
